chore: refactor issue domain model into dto's
